Fix EncFSVolume.movePath() for ChainedNameIV
There was a bug in EncFSVolume.movePath() that caused recursive moves
in the case of ChainedNameIV configurations to fail. Also improved
rollback handling in the case of an error during the recursive move
by trying to delete the newly created directories.

Added a FileNotFoundException to EncFSLocalFileProvider.move()
implementation as well.

+++ src/main/java/org/mrpdaemon/sec/encfs/EncFSLocalFileProvider.java
@@ -141,13 +141,17 @@ public class EncFSLocalFileProvider implements EncFSFileProvider {
 	 *            Path for the destination file or directory
 	 * 
 	 * @return true if the move is successful, false otherwise
-	 * 
+	 *
 	 * @throws IOException
 	 *             Source file/dir doesn't exist or misc. I/O error
 	 */
-	public boolean move(String srcPath, String dstPath) {
+	public boolean move(String srcPath, String dstPath) throws IOException {
 		File sourceFile = new File(rootPath.getAbsoluteFile(), srcPath);
 		File destFile = new File(rootPath.getAbsoluteFile(), dstPath);
+
+		if (!sourceFile.exists()) {
+			throw new FileNotFoundException("Path '" + srcPath + "' doesn't exist!");
+		}
 
 		return sourceFile.renameTo(destFile);
 	}
